kotlin trans
