+++ app/src/main/AndroidManifest.xml
@@ -1,5 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:tools="http://schemas.android.com/tools"
     package="com.saint.struct">
 
     <uses-permission android:name="android.permission.READ_EXTERNAL_STORAGE" />
@@ -7,21 +8,23 @@
     <uses-permission android:name="android.permission.INTERNET" />
     <uses-permission android:name="android.permission.ACCESS_NETWORK_STATE" />
     <uses-permission android:name="android.permission.WAKE_LOCK" />
-    <permission-group android:name="${applicationId}.andpermission"/>
+
+    <permission-group android:name="${applicationId}.andpermission" />
     <application
         android:name=".StructApp"
         android:allowBackup="true"
+        android:exported="false"
         android:icon="@mipmap/ic_launcher"
         android:label="@string/app_name"
         android:networkSecurityConfig="@xml/network_security_config"
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
-        android:exported = "false"
+        android:theme="@style/Theme.MaterialComponents.DayNight.NoActionBar"
+        tools:replace="android:theme"
         android:usesCleartextTraffic="true">
         <activity
             android:name=".activity.MainActivity"
-            android:exported="true"
-            android:theme="@style/AppTheme.NoActionBar">
+            android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
 
@@ -29,35 +32,21 @@
             </intent-filter>
         </activity>
 
-        <activity
-            android:name=".activity.WebActivity"
-            android:theme="@style/AppTheme.NoActionBar" />
+        <activity android:name=".activity.WebActivity" />
 
-        <activity
-            android:name=".activity.AidlActivity"
-            android:theme="@style/AppTheme.NoActionBar" />
-
-        <activity
-            android:name=".activity.PageOldActivity"
-            android:theme="@style/AppTheme.NoActionBar" />
-
-        <activity
-            android:name=".activity.PageKtActivity"
-            android:theme="@style/AppTheme.NoActionBar" />
-
-        <activity
-            android:name=".activity.PageRxActivity"
-            android:theme="@style/AppTheme.NoActionBar" />
+        <activity android:name=".activity.AidlActivity" />
 
         <service
             android:name=".task.service.MessengerService"
-            android:exported="false" />
+            android:exported="false"
+            tools:ignore="MissingClass" />
 
         <service
             android:name=".task.service.RemoteService"
             android:enabled="true"
             android:exported="false"
-            android:process=":other">
+            android:process=":other"
+            tools:ignore="MissingClass">
             <intent-filter>
                 <action android:name="com.saint.struct.aidl.Server.Action" />
             </intent-filter>
@@ -65,7 +54,8 @@
 
         <service
             android:name=".task.service.JobTestService"
-            android:permission="android.permission.BIND_JOB_SERVICE" />
+            android:permission="android.permission.BIND_JOB_SERVICE"
+            tools:ignore="MissingClass" />
 
         <meta-data
             android:name="com.google.android.actions"
