+++ app/src/main/java/com/saint/struct/adapter/Paging3KotlinAdapter.kt
@@ -1,17 +1,17 @@
 package com.saint.struct.adapter
 
-import android.annotation.SuppressLint
-import android.view.LayoutInflater
-import android.view.View
+import androidx.paging.PagingDataAdapter
 import android.view.ViewGroup
+import android.view.LayoutInflater
+import com.saint.struct.R
+import android.annotation.SuppressLint
+import android.view.View
 import android.widget.TextView
-import androidx.paging.PagedListAdapter
 import androidx.recyclerview.widget.DiffUtil
 import androidx.recyclerview.widget.RecyclerView
-import com.saint.struct.R
-import com.saint.struct.bean.WanListBean
+import com.saint.struct.bean.Project
 
-class PagingOldAdapter : PagedListAdapter<WanListBean, PagingOldAdapter.ViewHolder>(itemCallback) {
+class Paging3KotlinAdapter : PagingDataAdapter<Project, Paging3KotlinAdapter.ViewHolder>(itemCallback) {
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): ViewHolder {
         val view = LayoutInflater.from(parent.context).inflate(R.layout.item_page_recycle, parent, false)
         return ViewHolder(view)
@@ -29,26 +29,26 @@ class PagingOldAdapter : PagedListAdapter<WanListBean, PagingOldAdapter.ViewHold
 
     inner class ViewHolder(itemView: View) : RecyclerView.ViewHolder(itemView) {
         var desc: TextView
-        var date: TextView
         var title: TextView
+        var date: TextView
         var auth: TextView
 
         init {
+            title = itemView.findViewById(R.id.title)
             desc = itemView.findViewById(R.id.desc)
             date = itemView.findViewById(R.id.date)
-            title = itemView.findViewById(R.id.title)
             auth = itemView.findViewById(R.id.author)
         }
     }
 
     companion object {
-        private val itemCallback: DiffUtil.ItemCallback<WanListBean> = object : DiffUtil.ItemCallback<WanListBean>() {
-            override fun areItemsTheSame(oldItem: WanListBean, newItem: WanListBean): Boolean {
+        private val itemCallback: DiffUtil.ItemCallback<Project> = object : DiffUtil.ItemCallback<Project>() {
+            override fun areItemsTheSame(oldItem: Project, newItem: Project): Boolean {
                 return oldItem.id == newItem.id
             }
 
             @SuppressLint("DiffUtilEquals")
-            override fun areContentsTheSame(oldItem: WanListBean, newItem: WanListBean): Boolean {
+            override fun areContentsTheSame(oldItem: Project, newItem: Project): Boolean {
                 return oldItem == newItem
             }
         }
