+++ app/src/main/java/com/saint/struct/adapter/PagingRxJavaAdapter.kt
@@ -1,6 +1,7 @@
 package com.saint.struct.adapter
 
 import androidx.paging.PagingDataAdapter
+import com.saint.struct.bean.WanListBean
 import android.view.ViewGroup
 import android.view.LayoutInflater
 import com.saint.struct.R
@@ -9,9 +10,8 @@ import android.view.View
 import android.widget.TextView
 import androidx.recyclerview.widget.DiffUtil
 import androidx.recyclerview.widget.RecyclerView
-import com.saint.struct.bean.Project
 
-class PagingKtAdapter : PagingDataAdapter<Project, PagingKtAdapter.ViewHolder>(itemCallback) {
+class PagingRxJavaAdapter : PagingDataAdapter<WanListBean, PagingRxJavaAdapter.ViewHolder>(itemCallback) {
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): ViewHolder {
         val view = LayoutInflater.from(parent.context).inflate(R.layout.item_page_recycle, parent, false)
         return ViewHolder(view)
@@ -29,26 +29,26 @@ class PagingKtAdapter : PagingDataAdapter<Project, PagingKtAdapter.ViewHolder>(i
 
     inner class ViewHolder(itemView: View) : RecyclerView.ViewHolder(itemView) {
         var desc: TextView
-        var title: TextView
         var date: TextView
+        var title: TextView
         var auth: TextView
 
         init {
-            title = itemView.findViewById(R.id.title)
             desc = itemView.findViewById(R.id.desc)
             date = itemView.findViewById(R.id.date)
+            title = itemView.findViewById(R.id.title)
             auth = itemView.findViewById(R.id.author)
         }
     }
 
     companion object {
-        private val itemCallback: DiffUtil.ItemCallback<Project> = object : DiffUtil.ItemCallback<Project>() {
-            override fun areItemsTheSame(oldItem: Project, newItem: Project): Boolean {
-                return oldItem.niceDate == newItem.niceDate
+        private val itemCallback: DiffUtil.ItemCallback<WanListBean> = object : DiffUtil.ItemCallback<WanListBean>() {
+            override fun areItemsTheSame(oldItem: WanListBean, newItem: WanListBean): Boolean {
+                return oldItem.id == newItem.id
             }
 
             @SuppressLint("DiffUtilEquals")
-            override fun areContentsTheSame(oldItem: Project, newItem: Project): Boolean {
+            override fun areContentsTheSame(oldItem: WanListBean, newItem: WanListBean): Boolean {
                 return oldItem == newItem
             }
         }
