+++ app/src/main/java/com/saint/struct/fragment/BaseFragment.kt
@@ -1,0 +1,28 @@
+package com.saint.struct.fragment
+
+import android.os.Bundle
+import androidx.fragment.app.Fragment
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.databinding.DataBindingUtil
+import androidx.viewbinding.ViewBinding
+
+
+/**
+ * A simple [Fragment] subclass.
+ * Use the [BaseFragment.newInstance] factory method to
+ * create an instance of this fragment.
+ */
+abstract class BaseFragment : Fragment() {
+    lateinit var fragmentBinding: ViewBinding
+
+    override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View? {
+        fragmentBinding = DataBindingUtil.inflate(inflater, getFragmentLayoutId(), null, false)
+        doInit()
+        return fragmentBinding.root
+    }
+
+    abstract fun getFragmentLayoutId(): Int
+    abstract fun doInit()
+}
