+++ app/src/main/java/com/saint/struct/fragment/BasePagingFragment.kt
@@ -1,0 +1,45 @@
+package com.saint.struct.fragment
+
+import androidx.paging.PagingDataAdapter
+import androidx.recyclerview.widget.DividerItemDecoration
+import androidx.recyclerview.widget.LinearLayoutManager
+import androidx.recyclerview.widget.RecyclerView
+import com.saint.struct.R
+import com.saint.struct.databinding.FragmentPagingBinding
+
+abstract class BasePagingFragment<T : Any,VH : RecyclerView.ViewHolder> : BaseFragment() {
+    lateinit var pageAdapter:PagingDataAdapter<T,VH>
+
+    override fun doInit() {
+        initTitle()
+        initRecyclerView()
+        setModelAndData()
+        setListener()
+    }
+
+    open fun initTitle() {
+    }
+
+    override fun getFragmentLayoutId(): Int {
+        return R.layout.fragment_paging
+    }
+
+    private fun initRecyclerView() {
+        pageAdapter = getCustomPageAdapter()
+        val layoutManager = LinearLayoutManager(activity)
+        layoutManager.orientation = LinearLayoutManager.VERTICAL
+        (fragmentBinding as FragmentPagingBinding).recyclerView.apply {
+            this.layoutManager = layoutManager
+            this.addItemDecoration(DividerItemDecoration(activity, DividerItemDecoration.VERTICAL))
+            this.adapter = pageAdapter
+        }
+    }
+
+    abstract fun getCustomPageAdapter(): PagingDataAdapter<T, VH>
+
+    abstract fun setModelAndData()
+    open fun setListener(){
+
+    }
+
+}
