+++ app/src/main/java/com/saint/struct/fragment/Paging3KotlinFragment.kt
@@ -1,0 +1,35 @@
+package com.saint.struct.fragment
+
+import androidx.fragment.app.viewModels
+import androidx.lifecycle.ViewModelProvider
+import androidx.lifecycle.lifecycleScope
+import androidx.paging.PagingDataAdapter
+import com.saint.struct.adapter.Paging3KotlinAdapter
+import com.saint.struct.bean.Project
+import com.saint.struct.databinding.FragmentPagingBinding
+import com.saint.struct.viewmodel.Page3KtViewModel
+import kotlinx.coroutines.flow.collectLatest
+
+private const val TAG = "Paging3KotlinFragment"
+
+class Paging3KotlinFragment : BasePagingFragment<Project, Paging3KotlinAdapter.ViewHolder>() {
+
+    private val viewModel: Page3KtViewModel by viewModels()
+//    val viewModel :ArticleViewModel by viewModels()
+
+    override fun setModelAndData() {
+        lifecycleScope.launchWhenCreated {
+            viewModel.projects.collectLatest {
+                pageAdapter?.submitData(lifecycle, it)
+            }
+        }
+    }
+
+    override fun getCustomPageAdapter(): PagingDataAdapter<Project, Paging3KotlinAdapter.ViewHolder> {
+        return Paging3KotlinAdapter()
+    }
+
+    override fun initTitle() {
+        (fragmentBinding as FragmentPagingBinding).layoutAppBar.titleBar.text = "Paging3KotlinFragment"
+    }
+}
