+++ app/src/main/res/drawable/ic_tab_home.xml
@@ -1,0 +1,5 @@
+<?xml version="1.0" encoding="utf-8"?>
+<selector xmlns:android="http://schemas.android.com/apk/res/android">
+    <item android:drawable="@mipmap/ic_home_selected" android:state_checked="true"/>
+    <item android:drawable="@mipmap/ic_home" android:state_checked="false"/>
+</selector>
