+++ app/src/main/res/drawable/ic_tab_profile.xml
@@ -1,0 +1,5 @@
+<?xml version="1.0" encoding="utf-8"?>
+<selector xmlns:android="http://schemas.android.com/apk/res/android">
+    <item android:drawable="@mipmap/ic_profile_h" android:state_checked="true"/>
+    <item android:drawable="@mipmap/ic_profile" android:state_checked="false"/>
+</selector>
