+++ app/src/main/res/layout/fragment_paging.xml
@@ -8,12 +8,9 @@
         android:layout_height="match_parent"
         android:orientation="vertical">
 
-        <com.google.android.material.appbar.MaterialToolbar
-            android:id="@+id/toolbar"
-            android:layout_width="match_parent"
-            android:layout_height="?android:attr/actionBarSize"
-            android:background="@color/app_colorPrimary"
-           />
+        <include
+            android:id="@+id/layout_app_bar"
+            layout="@layout/layout_app_bar" />
 
         <androidx.recyclerview.widget.RecyclerView
             android:id="@+id/recyclerView"
