+++ app/src/main/res/menu/menu_bottom_nav.xml
@@ -1,0 +1,20 @@
+<?xml version="1.0" encoding="utf-8"?>
+<menu xmlns:android="http://schemas.android.com/apk/res/android">
+    <item
+        android:id="@+id/mainFragment"
+        android:icon="@drawable/ic_tab_home"
+        android:title="首页" />
+    <item
+        android:id="@+id/pageOldFragment"
+        android:icon="@drawable/ic_tab_rent"
+        android:title="详情" />
+    <item
+        android:id="@+id/pageKtFragment"
+        android:icon="@drawable/ic_tab_post"
+        android:title="列表" />
+    <item
+        android:id="@+id/pageRxFragment"
+        android:icon="@drawable/ic_tab_profile"
+        android:title="我的" />
+
+</menu>
