+++ app/src/main/res/navigation/nav_bottom_nav.xml
@@ -1,0 +1,24 @@
+<?xml version="1.0" encoding="utf-8"?>
+<navigation xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    app:startDestination="@id/mainFragment">
+
+    <fragment
+        android:id="@+id/mainFragment"
+        android:name="com.saint.struct.fragment.MainFragment"
+        android:label="MainFragment" />
+    <fragment
+        android:id="@+id/pageOldFragment"
+        android:name="com.saint.struct.fragment.PagingJavaFragment"
+        android:label="PageOldFragment" />
+    <fragment
+        android:id="@+id/pageKtFragment"
+        android:name="com.saint.struct.fragment.Paging3KotlinFragment"
+        android:label="PageKtFragment" />
+    <fragment
+        android:id="@+id/pageRxFragment"
+        android:name="com.saint.struct.fragment.Paging3RxJavaFragment"
+        android:label="PageRxFragment" />
+</navigation>
