+++ app/src/main/res/values/colors.xml
@@ -15,4 +15,15 @@
     <color name="colorPrimary">#1983d1</color>
     <color name="colorPrimaryDark">#1673b8</color>
     <color name="white">#ffffff</color>
+    <!--Sunflower project-->
+    <color name="sunflower_black">#de000000</color>
+    <color name="sunflower_gray_50">#fafafa</color>
+    <color name="sunflower_gray_50_a600">#99fafafa</color>
+    <color name="sunflower_green_300">#6dc790</color>
+    <color name="sunflower_green_500">#49bb79</color>
+    <color name="sunflower_green_700">#005d2b</color>
+    <color name="sunflower_green_900">#1a231e</color>
+    <color name="sunflower_white">#deffffff</color>
+    <color name="sunflower_yellow_300">#f8f99f</color>
+    <color name="sunflower_yellow_500">#ffff63</color>
 </resources>
